Navigate user to app page
