+++ lib/utils/utils.dart
@@ -3,3 +3,4 @@ export 'formatters.dart';
 export 'helpers.dart';
 export 'theme.dart';
 export 'app_launch_handler.dart';
+export 'platform_option.dart';
